Correção da requisição de extrato na API (API não exige uma query com o parametro do id do usuário)

+++ app/src/main/java/com/example/appbanksantander_accenturetalentacquisition/Model/StatementModel.kt
@@ -14,7 +14,7 @@ class StatementModel {
     var date: String = ""
 
     @SerializedName("value")
-    var value: Int = 0
+    var value: Float = 0.0F
